Change crawl to use breadth first search.
site_graph supports the --depth [int] argument to create an output graph that searches [int]
nodes away from the starting link (including the first node). Thus, if 0 is passed the empty
graph will be returned, were as if 1 is passed the singleton graph will be returned. Passing
any negative value forces the search to continue until the graph is exhausted as with DFS.